requirements.txt
urllib3
tweepy
simplejson
nltk
numpy
pytest
coveralls
python-dateutil
PySocks
pyopenssl
ndg-httpsclient
pyasn1
requests_oauthlib
requests
